Allow for binary payloads.
CAInfo_t's payload field only supported strings (0-terminated char arrays).
Added a payloadLength field to allow for raw byte data and changed some code to handle this.

Note:
The payload array should still have`payloadLength+1` items and have a '\0' character value at `payload[payloadLength]`
This is for backwards compatibility and safety (in case the `payload` is treated as a string).